+++ python/sglang/srt/layers/moe/token_dispatcher/__init__.py
@@ -1,0 +1,4 @@
+from sglang.srt.layers.moe.token_dispatcher.base_dispatcher import BaseDispatcher, BaseDispatcherConfig, DispatchOutput, DispatchOutputFormat
+from sglang.srt.layers.moe.token_dispatcher.deepep import DeepEPConfig, DeepEPDispatcher, DeepEPNormalOutput, DeepEPLLOutput
+
+__all__ = ["BaseDispatcher", "BaseDispatcherConfig", "DispatchOutput", "DispatchOutputFormat", "DeepEPConfig", "DeepEPDispatcher", "DeepEPNormalOutput", "DeepEPLLOutput"]
